26/02/06 정재은 등록자 수정,..등등

+++ src/main/java/com/yedam/app/attach/service/AttachmentService.java
@@ -3,7 +3,7 @@ package com.yedam.app.attach.service;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface AttachmentService {
-	Long saveSingleFile(String tableCode, Long userCode, MultipartFile uploadFile);
+	Long saveSingleFile(String tableCode, Integer userCode, MultipartFile uploadFile);
 	
 	void deleteSingleFile(Long fileCode);
 }
